requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
typing-extensions>=4.9.0

# Google Calendar integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# Assignment 4 — Fine-tuning, RAG & Evaluation
# FAISS wheels are typically unavailable on newer Python versions (e.g. 3.13+).
# The retriever includes a NumPy fallback when FAISS is not installed.
faiss-cpu>=1.9.0; python_version < "3.13"
tiktoken>=0.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0


# HTTP client (used by auth module)
requests>=2.31.0

# Database & Caching
pymongo>=4.7.0
redis>=5.0.0

# Assignment 6 — Testing & Visualization
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.12.0
mongomock>=4.1.0
matplotlib>=3.8.0
